Integrated the userId into the purchase Workflow

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/ReservationEndpoint.java
@@ -6,6 +6,7 @@ import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.SeatDto;
 import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundException;
 import at.ac.tuwien.sepm.groupphase.backend.service.PurchaseService;
 import at.ac.tuwien.sepm.groupphase.backend.service.ReservationService;
+import at.ac.tuwien.sepm.groupphase.backend.service.impl.CustomUserDetailService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.slf4j.Logger;
@@ -13,13 +14,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.*;
 
 import java.lang.invoke.MethodHandles;
 import java.util.List;
@@ -30,36 +25,39 @@ public class ReservationEndpoint {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private ReservationService service;
     private PurchaseService purchaseService;
+    private CustomUserDetailService userService;
 
     @Autowired
-    public ReservationEndpoint(ReservationService reservationService, PurchaseService purchaseService) {
+    public ReservationEndpoint(ReservationService reservationService, PurchaseService purchaseService, CustomUserDetailService userService) {
         this.service = reservationService;
         this.purchaseService = purchaseService;
+        this.userService = userService;
     }
 
     @Secured("ROLE_USER")
     @GetMapping
     @Operation(summary = "Gets a list of all Reservations from that user", security = @SecurityRequirement(name = "apiKey"))
-    public List<ReservationDto> getReservations() {
+    public List<ReservationDto> getReservations(@RequestHeader("Authorization") String token) {
         LOGGER.info("GET /api/v1/reservation");
 
-        //TODO: acquire UserID
-        //TODO: use real UserID
-        Long userId = 1L;
+        Long userId = userService.getUserIdFromToken(token);
+        if (userId == null) {
+            return null;
+        }
 
-        List<ReservationDto> itemList = service.getReservationsOfUser(userId);
-        return itemList;
+        return service.getReservationsOfUser(userId);
     }
 
     @Secured("ROLE_USER")
     @GetMapping("/{reservationNr}")
     @Operation(summary = "Gets a list of all Reservations from that user", security = @SecurityRequirement(name = "apiKey"))
-    public ReservationDto getReservation(@PathVariable Long reservationNr) {
+    public ReservationDto getReservation(@PathVariable Long reservationNr, @RequestHeader("Authorization") String token) {
         LOGGER.info("GET /api/v1/reservation/{}", reservationNr);
 
-        //TODO: acquire UserID
-        //TODO: use real UserID
-        Long userId = 1L;
+        Long userId = userService.getUserIdFromToken(token);
+        if (userId == null) {
+            return null;
+        }
 
         ReservationDto reservation = service.getReservationOfUser(reservationNr, userId);
         if (reservation == null) {
@@ -67,19 +65,22 @@ public class ReservationEndpoint {
             LOGGER.warn("something went wrong");
             return null;
         }
+
         return reservation;
     }
 
     @Secured("ROLE_USER")
     @PostMapping
     @Operation(summary = "Create a new Reservation with the given seats", security = @SecurityRequirement(name = "apiKey"))
-    public ResponseEntity<Void> newReservation(@RequestBody List<SeatDto> seatDtoList) {
+    public ResponseEntity<Void> newReservation(@RequestBody List<SeatDto> seatDtoList, @RequestHeader("Authorization") String token) {
         LOGGER.info("Post /api/v1/reservation");
-        try {
-            //TODO: acquire UserID
-            //TODO: use real UserID
-            Long userId = 1L;
 
+        Long userId = userService.getUserIdFromToken(token);
+        if (userId == null) {
+            return null;
+        }
+
+        try {
             service.addReservation(seatDtoList, userId);
         } catch (NotFoundException e) {
             return ResponseEntity.noContent().build();
@@ -90,12 +91,13 @@ public class ReservationEndpoint {
     @Secured("ROLE_USER")
     @DeleteMapping("/{reservationNr}")
     @Operation(summary = "Removes a Reservation by reservationnumber", security = @SecurityRequirement(name = "apiKey"))
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long reservationNr) {
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long reservationNr, @RequestHeader("Authorization") String token) {
         LOGGER.info("Delete /api/v1/reservation/{}", reservationNr);
 
-        //TODO: acquire UserID
-        //TODO: use real UserID
-        Long userId = 1L;
+        Long userId = userService.getUserIdFromToken(token);
+        if (userId == null) {
+            return null;
+        }
 
         this.service.deleteReservation(reservationNr, userId);
         return ResponseEntity.noContent().build();
@@ -104,12 +106,13 @@ public class ReservationEndpoint {
     @Secured("ROLE_USER")
     @PostMapping("/{reservationNr}/purchase")
     @Operation(summary = "Purchases a given amount of tickets of a Reservation", security = @SecurityRequirement(name = "apiKey"))
-    public ResponseEntity<Void> buyReservation(@RequestBody PurchaseCreationDto purchaseCreationDto, @PathVariable Long reservationNr) {
+    public ResponseEntity<Void> buyReservation(@RequestBody PurchaseCreationDto purchaseCreationDto, @PathVariable Long reservationNr, @RequestHeader("Authorization") String token) {
         LOGGER.info("Post /api/v1/cart/purchase");
 
-        //TODO: acquire UserID
-        //TODO: use real UserID
-        Long userId = 1L;
+        Long userId = userService.getUserIdFromToken(token);
+        if (userId == null) {
+            return null;
+        }
 
         purchaseService.purchaseReservationOfUser(reservationNr, purchaseCreationDto, userId);
         return ResponseEntity.ok().build();
